user/user/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
click==8.0.1
Flask==2.0.1
flask-restx==0.5.0
bcrypt==3.2.0
Flask-SQLAlchemy==2.5.1
python-dateutil==2.8.2
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
pyrsistent==0.18.0
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1
